requirements.txt
cfn_resource_provider
boto3


